+++ wajha/swift_theme/doctype/swift_home_card/swift_home_card.py
@@ -1,8 +1,0 @@
-# Copyright (c) 2026, Ali Raza and contributors
-# For license information, please see license.txt
-
-from frappe.model.document import Document
-
-
-class SwiftHomeCard(Document):
-	pass