+++ wajha/swift_theme/doctype/swift_theme_settings/__init__.py
@@ -1,6 +1,0 @@
-# Copyright (c) 2025, its-alikhokher
-# License: MIT
-
-from .swift_theme_settings import SwiftThemeSettings
-
-__all__ = ["SwiftThemeSettings"]