+++ wajha/swift_theme/doctype/swift_theme_sound_event/__init__.py
@@ -1,6 +1,0 @@
-# Copyright (c) 2025, its-alikhokher
-# License: MIT
-
-from .swift_theme_sound_event import SwiftThemeSoundEvent
-
-__all__ = ["SwiftThemeSoundEvent"]